experience-pro/pfe-we-tekup/code/django/requirements.txt
Django==4.2.2
djangorestframework
markdown
django-filter
djoser
django-cors-headers
psycopg2-binary
pillow
stripe